// src/lib/api/generated/models/orderFulfillmentType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 *
 */
export type OrderFulfillmentType = typeof OrderFulfillmentType[keyof typeof OrderFulfillmentType];


export const OrderFulfillmentType = {
  pickup: 'pickup',
  delivery: 'delivery',
} as const;